+++ web/api/helpers/app-rating/graphql/get-app-rating.generated.ts
