+++ web/api/helpers/app-store.ts
@@ -1,8 +1,6 @@
-"use server";
 import { getLocalisedCategory } from "@/lib/categories";
 import { whitelistedAppsPermit2 } from "@/lib/constants";
 import { generateExternalNullifier } from "@/lib/hashing";
-import { createRedisClient } from "@/lib/redis";
 import {
   AppStatsReturnType,
   AppStoreFormattedFields,
@@ -10,67 +8,7 @@ import {
   AppStoreMetadataFields,
 } from "@/lib/types";
 import { getCDNImageUrl, tryParseJSON } from "@/lib/utils";
-import { getAPIServiceGraphqlClient } from "../graphql";
-import { getSdk as getAppRatingSdk } from "./graphql/get-app-rating.generated";
-
-// Helper function to get rating with Redis caching
-export async function getAppRating(appId: string): Promise<number> {
-  const redisKey = `app:${appId}:rating`;
-  const lockKey = `lock:${appId}:rating`;
-
-  const redis = createRedisClient({
-    url: process.env.REDIS_URL!,
-    password: process.env.REDIS_PASSWORD!,
-    username: process.env.REDIS_USERNAME!,
-  });
-
-  try {
-    // Try to get from cache first
-    let rating = await redis.get(redisKey);
-
-    if (rating !== null) {
-      return parseFloat(rating);
-    }
-
-    // Try to acquire lock
-    const acquiredLock = await redis.set(lockKey, "pending", "EX", 30);
-
-    if (!acquiredLock) {
-      for (let i = 0; i < 3; i++) {
-        await new Promise((resolve) => setTimeout(resolve, 1000));
-        rating = await redis.get(redisKey);
-        if (rating !== null) {
-          return parseFloat(rating);
-        }
-      }
-
-      console.warn("Lock timeout for app rating calculation", { appId });
-    }
-    const client = await getAPIServiceGraphqlClient();
-
-    // Calculate rating from DB
-    const result = await getAppRatingSdk(client).GetAppRating({
-      app_id: appId,
-    });
-
-    const calculatedRating = result.app_metadata[0].app_rating ?? 0;
-
-    // Cache for 24 hours
-    await redis.set(redisKey, calculatedRating.toString(), "EX", 24 * 60 * 60);
-
-    return calculatedRating; // Return as float
-  } catch (error) {
-    console.warn("Error getting app rating with cache", { error, appId });
-    return 0; // Return 0 if there's an error
-  } finally {
-    try {
-      await redis.del(lockKey);
-      await redis.quit();
-    } catch (redisError) {
-      console.error("Error closing Redis connection", { redisError });
-    }
-  }
-}
+import { getAppRating } from "./app-rating";
 
 export const formatAppMetadata = async (
   appData: AppStoreMetadataFields,
@@ -83,6 +21,7 @@ export const formatAppMetadata = async (
     0;
 
   const appRating = await getAppRating(appMetadata.app_id);
+  console.log("appRating", appRating);
 
   const localisedContent = appMetadata.localisations?.[0];
