+++ web/tests/api/v2/public/apps/apps.test.ts
@@ -43,8 +43,15 @@ jest.mock(
   }),
 );
 
+jest.mock("@/api/helpers/app-rating", () => ({
+  getAppRating: jest.fn().mockResolvedValue(3.4),
+}));
+
 beforeEach(() => {
   jest.resetAllMocks();
+
+  const { getAppRating } = require("@/api/helpers/app-rating");
+  getAppRating.mockResolvedValue(3.4);
 });
 
 describe("/api/v2/public/apps", () => {
@@ -224,6 +231,7 @@ describe("/api/v2/public/apps", () => {
             },
             world_app_button_text: "random",
             world_app_description: "random",
+            app_rating: 3.4,
             verification_status: "verified",
           },
           {
@@ -253,6 +261,7 @@ describe("/api/v2/public/apps", () => {
             team_name: "Example Team",
             world_app_button_text: "random",
             world_app_description: "random",
+            app_rating: 3.4,
             verification_status: "verified",
           },
           {
@@ -270,6 +279,7 @@ describe("/api/v2/public/apps", () => {
             ],
             whitelisted_addresses: ["0x1234", "0x5678"],
             unique_users: 0,
+            app_rating: 3.4,
             app_mode: "mini-app",
             associated_domains: ["https://worldcoin.org"],
             contracts: ["0x0c892815f0B058E69987920A23FBb33c834289cf"],
@@ -330,7 +340,6 @@ describe("/api/v2/public/apps", () => {
             permit2_tokens: ["0x0c892815f0B058E69987920A23FBb33c834289cf"],
             supported_countries: ["us"],
             supported_languages: ["en", "es"],
-            app_rating: 3.4,
             verification_status: "verified",
             app: {
               team: {
